Use stable version for VirtualBox on host system
In order to be able to test VirtualBox in a Docker container with the
help of Ansible Molecule, the VirtualBox kernel driver must be loaded on
the host system. For this purpose, the latest stable version of this
Ansible role is first downloaded under Travis CI/CD via Ansible
Galaxy, and then VirtualBox including the VirtualBox kernel driver is
installed on the host system. The new device /dev/vboxdrv is created
on the host system, which is then mounted in the respective Docker image
with the help of Ansible Molecule. With the help of this Docker image,
the new version of this Ansible role can then be tested.